requirements.txt
phidata
python-dotenv
yfinance>=0.2.0
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
phi-agent>=0.1.0